refactor(http): [#160] reorganize dirs for Axum and Warp HTTP tracker implementations
We are going to start sharing code bettween both implementation (Warp
and Axum). We need to keep common code separated because Warp
implementation will be removed when Axum implementation is finished.
