+++ src/http/warp_implementation/filters.rs
@@ -7,8 +7,8 @@ use std::sync::Arc;
 use warp::{reject, Filter, Rejection};
 
 use super::error::Error;
-use super::percent_encoding::{percent_decode_info_hash, percent_decode_peer_id};
 use super::{request, WebResult};
+use crate::http::percent_encoding::{percent_decode_info_hash, percent_decode_peer_id};
 use crate::protocol::common::MAX_SCRAPE_TORRENTS;
 use crate::protocol::info_hash::InfoHash;
 use crate::tracker::{self, auth, peer};
