+++ src/jobs/http_tracker.rs
@@ -7,8 +7,8 @@ use tokio::sync::oneshot;
 use tokio::task::JoinHandle;
 
 use crate::config::HttpTracker;
-use crate::http::axum::server;
-use crate::http::server::Http;
+use crate::http::axum_implementation::server;
+use crate::http::warp_implementation::server::Http;
 use crate::http::Version;
 use crate::tracker;
